* Call a shell script to run the ping utility rather than hardcoding it in code. The script should be called slbd.sh and placed on /usr/local/sbin. There is one default supplied with the port. * Fix some bound checks which needed to be for INT but were for LONG * Fix poller_ping function signature/declaration. The port builds without any warnings now.

+++ pfPorts/slbd/files/slbd.c
@@ -95,7 +95,7 @@ int main(int argc, char **argv) {
 			break;
 		case 'r':
 			r_refresh = strtol(optarg, NULL, 10);
-			if (r_refresh == LONG_MIN || r_refresh == LONG_MAX)
+			if (r_refresh == INT_MIN || r_refresh == INT_MAX)
 				err(1, "Bad refresh time");
 			break;
 		default:
